Update stress-tests for tryorama 0.2.0
