+++ stress-test/src/telephone-games.ts
@@ -1,17 +1,7 @@
-import { Config } from '@holochain/tryorama'
 import * as R from 'ramda'
+import { Config } from '@holochain/tryorama'
+import { configBatchSimple } from './config'
 
-const dna = Config.dna('passthrough-dna.dna.json', 'passthrough')
-
-/** Generates a bunch of identical conductor configs with multiple identical instances */
-const configBatchSimple = (numConductors, numInstances) => {
-    const conductor = R.pipe(
-        R.map(n => [`${n}`, dna]),
-        R.fromPairs,
-        x => ({ instances: x }),
-    )(R.range(0, numInstances))
-    return R.repeat(conductor, numConductors)
-}
 
 // This is a customizable blue print for "telephone game" network/time topology setups.
 // It cycles through all N agents in a way that at any given point only two agents
